requirements.txt
google-adk
requests
tavily-python
google-cloud-aiplatform